pyproject.toml
[project]
name = "courses"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13.5"
dependencies = [
    "cookiecutter>=2.6.0",
    "django>=5.2.4",
    "django-browser-reload>=1.18.0",
    "django-mptt>=0.17.0",
    "django-tailwind>=4.2.0",
    "nodeenv>=1.9.1",
    "pillow>=11.3.0",
    "psycopg2-binary>=2.9.10",
    "python-slugify>=8.0.4",
]
